added room database for local data storage

+++ app/src/main/java/com/framework/openweatherandroidapp/model/Main.java
@@ -11,10 +11,10 @@ public class Main {
     private double tempMin;
 
     @SerializedName("humidity")
-    private int humidity;
+    private double humidity;
 
     @SerializedName("pressure")
-    private int pressure;
+    private double pressure;
 
     @SerializedName("temp_max")
     private double tempMax;
@@ -35,19 +35,19 @@ public class Main {
         return tempMin;
     }
 
-    public void setHumidity(int humidity) {
+    public void setHumidity(double humidity) {
         this.humidity = humidity;
     }
 
-    public int getHumidity() {
+    public double getHumidity() {
         return humidity;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(double pressure) {
         this.pressure = pressure;
     }
 
-    public int getPressure() {
+    public double getPressure() {
         return pressure;
     }
